% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFATable.R
\name{CFATable}
\alias{CFATable}
\title{CFATable}
\usage{
CFATable(
  CFAfit,
  header,
  digits = 3,
  footer,
  adjtopage = c("portrait", F, 1, 3),
  CorMatrix = F,
  FCorMat = NULL,
  savename
)
}
\arguments{
\item{CFAfit}{A lavaan cfa fit object.}

\item{header}{Name of the table in quotes.}

\item{digits}{Number of decimal places in the table. By default digits = 3.}

\item{footer}{Optional argument. Inserts a general table footnote. The argument must be in quotes.}

\item{adjtopage}{Set columns width to default page margins (2.5) and set space between text and lines.This argument must be a vector composed of: orientation of page ("landscape/portrait"), does first column need to be wider (True/False), if previews element set to TRUE set ratio of first column size (numeric greater than 0), space between lines and text (numeric, in points). Default is c("portrait", F,1,3).}

\item{CorMatrix}{Include the factor correlation matrix in the table. Default = FALSE.}

\item{FCorMat}{Include a specific factor correlation matrix (for exemple after factor score computation). If FCorMat=NULL and CorMatrix=TRUE, psi matrix from CFAfit.}

\item{savename}{Optional argument. Saves the table in .docx format in the working space. The argument must be in quotes.}
}
\value{
CFA APA Table
}
\description{
Create an APA table from lavaan cfa fit object.
}
\examples{

cfamodel <- '
  f1 ~= V1 + V2 + V3
  f2 ~= V4 + V5 + V6
'

cfafit <- cfa(model = cfamodel, data = df)

CFATable(cfafit, "CFA APA table")
}
\author{
Simon Pineau
}
