% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiTable.R
\name{ChiTable}
\alias{ChiTable}
\title{ChiTable}
\usage{
ChiTable(
  chitest,
  header,
  digits = 2,
  expected = FALSE,
  footer,
  adjtopage,
  savename
)
}
\arguments{
\item{chitest}{A chi² object from chisq.test function}

\item{header}{Name of the table in quotes.}

\item{digits}{Number of decimal places in the table. By default digits = 2.}

\item{expected}{Does the table need to include expected repartition. Default = FALSE.}

\item{footer}{Optional argument. Inserts a general table footnote. The argument must be in quotes.}

\item{adjtopage}{Set columns width to default page margins (2.5) and set space between text and lines.This argument must be a vector composed of: orientation of page ("landscape/portrait"), does first column need to be wider (True/False), if previews element set to TRUE set ratio of first column size (numeric greater than 0), space between lines and text (numeric, in points). Default = adjtopage = c("portrait", F, 1, 3).}

\item{savename}{Optional argument. Saves the table in .docx format in the working space. The argument must be in quotes.}
}
\value{
APA Table
}
\description{
Create an APA norme chi² table.
}
\examples{

chi2<-chisq.test(table(df[c("V1","V2")]))

ChiTable(chi2, header = "chi2 table", expected = F, adjtopage = c("portrait", F, 1, 3))
}
\author{
Simon Pineau
}
