% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableFun.R
\name{TableFun}
\alias{TableFun}
\title{TableFun}
\usage{
TableFun(
  df,
  header,
  digits = 2,
  left.align = NULL,
  right.align = NULL,
  top.align = NULL,
  bottom.align = NULL,
  merge.col = 1,
  footer = NULL,
  foot.note.header = NULL,
  foot.note.body = NULL,
  savename = NULL
)
}
\arguments{
\item{df}{Un jeu de données (dataframe).}

\item{header}{Nom du tableau entre guillemets.}

\item{digits}{Nombre de décimale dans le tableau. Par default digits = 2.}

\item{left.align}{Argument optionnel. Aligne les informations du corps du tableau à gauche.}

\item{right.align}{Argument optionnel. Aligne les informations du corps du tableau à droite.}

\item{top.align}{Argument optionnel. Aligne les informations du corps du tableau en haut.}

\item{bottom.align}{Argument optionnel. Aligne les informations du corps du tableau en bas.}

\item{merge.col}{Funsionne les lignes similaires adjacentes des colonnes spécifiées. Par default merge.col = 1.}

\item{footer}{Argument optionnel. Insert une note de bas de tableau générale. L'argument doit être entre guillemets.}

\item{foot.note.header}{Argument optionnel. Insert une note specifique attachée un titre de colonne. Cette argument doit être un vecteur composé : du numéro de la colonne, de la "lettre" associé à la note, et de la "note specifique" : c(1,"a","note specifique"). Plusieurs notes spécifiques peuvent être ajoutées : c(1,"a","note specifique1", 2,"b","note specifique2").}

\item{foot.note.body}{Argument optionnel. Insert une note specifique attachée à une cellule du corps du tableau. La rédaction de l'argument est la même que celle de foot.note.header en ajoutant le numéro de la ligne après le numéro de la colonne : c(1, 1, "a","note specifique")}

\item{savename}{Argument optionnel. Enregistre le tableau au format .docx dans l'esapce de travail. L'argument doit être entre guillemets.}
}
\description{
Permet de créer des tableaux aux norme APA à partir de dataframe.
}
\author{
Simon Pineau
}
